FastCholesky

Documentation for FastCholesky.

FastCholesky.cholinv

FastCholesky.cholinv_logdet

FastCholesky.chollogdet

FastCholesky.cholsqrt

FastCholesky.fastcholesky

FastCholesky.fastcholesky!

FastCholesky.cholinv — Method

cholinv(input)

Returns an inverse of the input, but uses the Cholesky factorization. An alias for inv(fastcholesky(input)).

FastCholesky.cholinv_logdet — Method

cholinv_logdet(input)

Returns an inverse and a log-determinant of the matrix input simultaneously, but uses the Cholesky factorization.

FastCholesky.chollogdet — Method

cholsqrt(input)

Returns a log-determinant of the matrix input, but uses the Cholesky factorization. An alias for logdet(fastcholesky(input)).

FastCholesky.cholsqrt — Method

cholsqrt(input)

Returns an squared root of the matrix input, but uses the Cholesky factorization. An alias for fastcholesky(x).L.

FastCholesky.fastcholesky! — Function

fastcholesky!(input)

In-place version of the fastcholesky. Does not check the positive-definiteness of the input and does not throw. Use LinearAlgebra.issuccess to check if the result is a proper Cholesky factorization.

FastCholesky.fastcholesky — Method

fastcholesky(input)

Returns the Cholesky factorization of the input in the same format as LinearAlgebra.cholesky. By default fallbacks to LinearAlgebra.cholesky(PositiveFactorizations.Positive, input), thus does not requires the input to be perfectly symmetric. Provides more efficient implementations for certain inputs.